+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-Bne4sD1z.js
@@ -1,8 +1,0 @@
-import{c as r,s as e}from"./flowDiagram-HODETNUW-CLYpb995.js";import{_ as a}from"./mermaid.core-bKLjiR6p.js";import"./chunk-5VM5RSS4-BKIcICGV.js";import"./chunk-XXDRQBXY-VvxgKztM.js";import"./chunk-POPQ4Y6H-B_OvUkEf.js";import"./chunk-F27PBJKO-CYeu61sg.js";import"./channel-DZONH1aQ.js";import"./index-JaIbObHs.js";var o=a(t=>`${e(t)}
-  .swimlane.cluster rect {
-    stroke: ${t.clusterBorder} !important;
-  }
-  [data-look="neo"].cluster rect {
-    filter: none;
-  }
-`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};