+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-C8Cfw7DL.js
@@ -1,0 +1,8 @@
+import{c as r,s as e}from"./flowDiagram-HODETNUW-B41Fbqtx.js";import{_ as a}from"./mermaid.core-DY3nXMcK.js";import"./chunk-5VM5RSS4-CfRbRdvy.js";import"./chunk-XXDRQBXY-rl4C_M9C.js";import"./chunk-POPQ4Y6H-CtVRIf0S.js";import"./chunk-F27PBJKO-B2FAX8oX.js";import"./channel-BqGrPzs6.js";import"./index-BrbluF_t.js";var o=a(t=>`${e(t)}
+  .swimlane.cluster rect {
+    stroke: ${t.clusterBorder} !important;
+  }
+  [data-look="neo"].cluster rect {
+    filter: none;
+  }
+`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};